+++ CHANGES.rst
@@ -6,9 +6,12 @@ Changelog
 
 - Add get_environment method and test.
   [bsuttor]
-
 - is_develop_environment method is true if global environment variable 'ENV' is equal to 'dev'.
   [bsuttor]
+- Added 'cache' module with helper methods 'cleanVocabularyCacheFor' that will clean
+  instance.memoize cache defined on a named vocabulary and 'cleanRamCacheFor' that
+  will clean ram.cache defined on a given method
+  [gbastien]
 
 
 0.4.9 (2015-04-21)